Code cleanup to avoid redunancy and eliminate single use routines
