+++ src/axom/primal/tests/primal_clip.cpp
@@ -11,6 +11,7 @@
 #include "axom/core/execution/for_all.hpp"
 #include "axom/core/memory_management.hpp"
 #include "axom/core/numerics/transforms.hpp"
+#include "axom/core/utilities/MemoryTesting.hpp"
 
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/geometry/BoundingBox.hpp"
@@ -42,18 +43,6 @@ using PolygonType = axom::primal::Polygon<double, 3>;
 using PlaneType = axom::primal::Plane<double, 3>;
 using PolyhedronType = axom::primal::Polyhedron<double, 3>;
 }  // namespace Primal3D
-
-template <typename ExecSpace>
-int globalDefaultAllocatorForExecSpace()
-{
-#if defined(AXOM_USE_UMPIRE)
-  return axom::execution_space<ExecSpace>::onDevice()
-    ? axom::execution_space<ExecSpace>::allocatorID()
-    : axom::getUmpireResourceAllocatorID(umpire::resource::Host);
-#else
-  return axom::execution_space<ExecSpace>::allocatorID();
-#endif
-}
 
 /// Compare two polygons with a tolerance.
 template <typename PolygonType, typename T>
@@ -345,7 +334,7 @@ void unit_check_poly_clip()
   // In addition, vertices 0 and 3 should be marked as clipped.
 
   const int current_allocator = axom::getDefaultAllocatorID();
-  axom::setDefaultAllocator(globalDefaultAllocatorForExecSpace<ExecPolicy>());
+  axom::setDefaultAllocator(axom::utilities::globalDefaultAllocatorForExecSpace<ExecPolicy>());
 
   PolyhedronType* out_square = axom::allocate<PolyhedronType>(1);
   unsigned int* out_clipped = axom::allocate<unsigned int>(1);
@@ -443,7 +432,7 @@ void check_hex_tet_clip(double EPS)
   const int current_allocator = axom::getDefaultAllocatorID();
 
   // Set new default to device if available
-  axom::setDefaultAllocator(globalDefaultAllocatorForExecSpace<ExecPolicy>());
+  axom::setDefaultAllocator(axom::utilities::globalDefaultAllocatorForExecSpace<ExecPolicy>());
 
   // Allocate memory for shapes
   TetrahedronType* tet = axom::allocate<TetrahedronType>(1);
@@ -522,7 +511,7 @@ void check_oct_tet_clip(double EPS)
   const int current_allocator = axom::getDefaultAllocatorID();
 
   // Set new default to device if available
-  axom::setDefaultAllocator(globalDefaultAllocatorForExecSpace<ExecPolicy>());
+  axom::setDefaultAllocator(axom::utilities::globalDefaultAllocatorForExecSpace<ExecPolicy>());
 
   // Allocate memory for shapes
   TetrahedronType* tet = axom::allocate<TetrahedronType>(1);
@@ -598,7 +587,7 @@ void check_tet_tet_clip(double EPS)
   const int current_allocator = axom::getDefaultAllocatorID();
 
   // Set new default to device if available
-  axom::setDefaultAllocator(globalDefaultAllocatorForExecSpace<ExecPolicy>());
+  axom::setDefaultAllocator(axom::utilities::globalDefaultAllocatorForExecSpace<ExecPolicy>());
 
   // Allocate memory for shapes
   TetrahedronType* tet1 = axom::allocate<TetrahedronType>(1);
